CHANGED complete rework of the net code

+++ main/java/com/gamepad/lib/net/IpAddress.java
@@ -7,10 +7,10 @@ import java.util.ArrayList;
  */
 public class IpAddress
 {
-    private Byte[] ipAddress;
+    private Integer[] ipAddress;
     //commit change
 
-    public IpAddress(Byte[] ipAddress)
+    public IpAddress(Integer[] ipAddress)
     {
         this.ipAddress = ipAddress;
     }
@@ -29,7 +29,7 @@ public class IpAddress
         return ipAddress[0] + "." + ipAddress[1] + "." + ipAddress[2] + "." + ipAddress[3];
     }
 
-    public Byte[] getBytes()
+    public Integer[] getBytes()
     {
         return ipAddress;
     }
@@ -57,16 +57,16 @@ public class IpAddress
         {
             return false;
         }
-        ArrayList<Byte> ipPartList = new ArrayList<Byte>();
+        ArrayList<Integer> ipPartList = new ArrayList<Integer>();
         for(String ipPart : ipParts)
         {
             if(ipPart.equals(""))
             {
                 return false;
             }
-            ipPartList.add(Byte.parseByte(ipPart));
+            ipPartList.add(Integer.parseInt(ipPart));
         }
-        objectToSaveIn = new IpAddress(ipPartList.toArray(new Byte[0]));
+        objectToSaveIn = new IpAddress(ipPartList.toArray(new Integer[0]));
         return true;
     }
